moved DNSUtils methods to submodule level
